requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
requests==2.31.0
newsapi-python==0.2.7
gunicorn==21.2.0